requirements.txt
fastapi
uvicorn
openai
faiss-cpu
sentence-transformers
pdfplumber
python-dotenv
